Remove AXTreeSnapshotter stress test experiments on Android
This experiment ran in Q1 2024, and it gave enough data to be useful,
and we do not need to run the experiment again. With this CL we remove
all the lingering code that enabled the experiment.

OBSOLETE_HISTOGRAM[Accessibility.AXTreeSnapshotter.Snapshot.EndToEndRuntime]=This was a histogram for experimental data collection, and no longer serves any purpose.
OBSOLETE_HISTOGRAM[Accessibility.AXTreeSnapshotter.Snapshot.NoRestrictions.Time]=This was a histogram for experimental data collection, and no longer serves any purpose.
OBSOLETE_HISTOGRAM[Accessibility.AXTreeSnapshotter.Snapshot.NoRestrictions.Nodes]=This was a histogram for experimental data collection, and no longer serves any purpose.

AX-Relnotes: N/A
Bug: N/A
Change-Id: I01dafe951d8f34a76e69dcebe7acd81d2b4f6107
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6092352
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1397997}