[PWA] Test for not reparenting for install from app window.
This change 'pulls up' the check for the current web contents being in
an app window to the WebAppUiManager, so that other callers
(pwa_handler.cc) can also ensure they do not reparent incorrectly.

This change also adds a test for this protection.

Fixed: 384549647, 383360269
Change-Id: I0f03ba36cad282c31b37afaa4c943f3fc34e3b4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6096698
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398001}