Enable (by default) audio policy flag check in Android.
hardware.audio_capture_enabled will be used to gate whether the
mic action will be shown in the Android UI surfaces.

Added in M88, but never enabled. This will enable by default but
remain in the code (for a milestone) if needed to disable via a
server kill switch.

Bug: 40738542
Change-Id: Ic72c2d56ca8e83240a2bc95de7bbe64780baaef3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6103140
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Reviewed-by: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1398024}