[lensoverlay] Fix failing Mac tests
The tests were crashing after this[1] fieldtrial config that caused a race condition which would cause the initialization to not be ready when the objects response came in. This primarily happened in the PDF tests since the PDF tests slow down the initialization time. To remedy this, I
created a new variable to store the data that comes in prior to the initialization data being ready.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/6082871

Bug: 384549641, 383554290
Change-Id: Ic9945f3329b2da6b9ca514a4304aed337f78a7a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6102410
Commit-Queue: Duncan Mercer <mercerd@google.com>
Reviewed-by: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1398044}