Improve Autofill on typing acceptance metrics
Today this metric has the following problem, in a scenario where:

1. The user accepts a suggestion for field A and navigates away
2. The user navigates back fo field A (same global id) and does not
   accept a suggestion.

We are only going to emit a metric where the suggestion was not
accepted, effectively ignoring the first acceptance.

This CL effectively emits acceptance metric on the stop, the non
accepted bucket is only emitted in destruction time if no suggestion
was ever accepted on that field. With the new implementation the above
scenario will correctly emit two counts, one where the suggestion was
not accepted and one where it was.


Bug: 381994105
Change-Id: Ia7a34f578add25f318100e986feb02e233e12bb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6097803
Auto-Submit: Bruno Braga <brunobraga@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1398082}