[Tab Groups] Account for large text size in tab group dialog
Fix two row sizing for large font sizes.

Before this fix the text in the share button and for the edit text
would be clipped. After this fix, the share button resizes
automatically to fit the text and the EditText padding is removed to
fix the alignment (which looks better).

This does cause some subtle change to the rendering both before and
after since it is easier to apply the padding/changes consistently
than to special case based on text size.

Demos:
https://drive.google.com/drive/folders/1DBSG505X0PTH9edRc7vdYmYS0MhfZW4Z?usp=sharing

Fixed: 382440347
Change-Id: Ibbf7d0a1135eea599eeab9951005d8b10931375e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6104218
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398086}