[ntp] Support NTP loads in app contexts
Currently it is possible for NTP loads to occur in app contexts.

Consequently it is possible for NTP loads to start without a hosting
TabModel, and for the NTP to be moved between window types (normal<>app
browser transitions).

This CL ensures the NTP will not crash and correctly function in both
of the above mentioned cases. The CustomizeChromeSidePanelController
tab feature has also been transitioned to leverage the new WebUI
embedding context utils.

Handling for these the above cases of absent and changing TabModels
can be removed once the NTP is disallowed from loading in app
browsers.

Bug: 378475391
Change-Id: Ibf466cf2ee29e8f2f1ec4c1f5367470a6bd4dbcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6040425
Reviewed-by: Roman Arora <romanarora@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398120}