Change Observers: Make FileSystemChangeRecord a dictionary
It is still being discussed on the spec PR whether the
FileSystemChangeRecord should be a dictionary or an interface. To leave
our options open, this switches it to a dictionary for now since we can
switch to an interface without breaking anything but not vice versa.

Fixed: 378924423
Change-Id: I73427d32b132cdf4a0af518be775220b3c0792b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6099067
Reviewed-by: Daseul Lee <dslee@chromium.org>
Commit-Queue: Daseul Lee <dslee@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Auto-Submit: Nathan Memmott <memmott@chromium.org>
Commit-Queue: Ayu Ishii <ayui@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398122}