[IOS] Fix IOS.whatsNew.{Type}.timeSpent not recorded on back button
Issue: The IOS.whatsNew.{Type}.timeSpent was only logged when the
"Done" button was tapped, not when the "Back" button was used.

Change: Updated the stop method for whats_new_detail_coordinator to ensure logTimeSpentOnDetailView is not depending on the whatsNewScreenshotViewController being in the navigation stack when stopping the coordinator.

Fix: Time spent on the What's New details screen is now logged
correctly when either the "Done" or "Back" button is pressed.

Bug: 354738315

Screenshot of Fix IOS.whatsNew.{Type}.timeSpent Fix:
https://screenshot.googleplex.com/AESJLfKm38uPxfs

Change-Id: Ib8dbb1bb7b1328b500593e693141877a2ad34008
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6105120
Reviewed-by: Cheick Cisse <cheickcisse@google.com>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Ibrahim Kanouche <kanouche@google.com>
Cr-Commit-Position: refs/heads/main@{#1398127}