[PWA/NavCapture] Add 'enable for non-auto client modes' mode.
This change adds another mode for the navigation capturing feature where
navigation capturing is enabled for an app only when the app sets a non-auto client mode in it's manifest.

When we attempt to launch the on-by-default version, there is a chance
that users or devs are annoyed that new app windows open all the time
for apps that don't set a client_mode in the manifest (instead of
keeping one window open only). This setting will enable the navigation
capturing by default for all apps that set a client_mode in the
manifest, which will be much more directly correlated to a good user
experience as the dev has specified the client mode launch handling
behavior.

R=dibyapal@chromium.org

Fixed: 384797864
Change-Id: I3df5cff9d5134e9420292dcf59f7df5ce3c968d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6102391
Reviewed-by: Dibyajyoti Pal <dibyapal@chromium.org>
Auto-Submit: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398188}