Remove chromium-accessibility@chromium.org as owners of histograms
Set chromeos-a11y@google.com as the secondary owner instead.

This is a Google group with over a thousand public users, it shouldn't be listed as an owner. When an expiry bug crbug.com/364018740 was filed for one of the histograms, the public group was cc'd and consequently gave public users access to the bug.

Change-Id: I32ecf8ac54cd51f8f4278eafbbeca69eacaaca47
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6104647
Reviewed-by: Katie Dektar <katie@chromium.org>
Reviewed-by: Sun Yueru <yrsun@chromium.org>
Commit-Queue: Ariel Zhang <arielzhang@google.com>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Cr-Commit-Position: refs/heads/main@{#1398214}