Update menus on bookmark model changes (3/N)
This CL adds an "OnBookmarkNodeMoved" handler to `BookmarkMenuDelegate`,  which updates menu items to reflect this. The high-level steps to achieve this are: remove the menu item of the moved node from the older parent (if built); update menu artifacts (e.g., bookmarks title, separators); build a new menu in the new parent node's menu (if built).

*Why not just rebuild the whole menu?*
1. Potentially high volume of bookmarks. Lazy-loading was initially implemented[1] to handle raising the bookmarks limit to 10,000.
2. Destroying + rebuilding nested menus creates a "flickering" effect.
3. For DnD, clients don't behave well when the widget running the drag loop is deleted mid-drag. This will be investigated as part of crbug.com/382711086.

[1]https://codereview.chromium.org/1005873012

Change-Id: Ib231fc6652523264b0f8f363c68dfddd806b9a0d
Bug: 40868588
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6084618
Reviewed-by: Dana Fried <dfried@chromium.org>
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398229}