[lensoverlay] Allow null TextLayout.
TextLayout is optional in the Lens response and is not filled if no text
is detected, but this was not reflected in the mojo interface and
resulted in an exception. This was causing the region selection context
menu to not show in some cases.

Bug: b:370825133
Change-Id: Ia7d1b875cd0fc0df7b49311e616fd8807b3e4434
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6107270
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: Bryan Nguyen <nguyenbryan@google.com>
Reviewed-by: Juan Mojica <juanmojica@google.com>
Cr-Commit-Position: refs/heads/main@{#1398284}