[A11y] Caret offset incorrect when caret inside embedded object
We have code that automatically increments the offset if it's on
an embedded object. It was added in 2015 in this CL:
https://codereview.chromium.org/1377733002.
It's now in AXPlatformNodeBase::GetSelectionOffsetsFromTree().
The comment says "The IA2 Spec says that if the largest of the two
offsets falls on an embedded object character and if there is a
selection in that embedded object, it should be incremented by one so
that it points after the embedded object character. This is a signal to
AT software that the embedded object is also part of the selection."

However, it's weird if this code is clearly wrongly applied to caret
offsets, because it makes it look like the caret is after the embedded
object when there's a selection inside. Some relevant text from the
IAcceessibleText docs at https://wiki.linuxfoundation.org/accessibility/ia2/ia2_implementation_guide :
"As mentioned in item B, in the section entitled “The caret”, the client must be able to start from the object with focus, and by recursively using the IAccessibleHypertext interface, find the object that contains the caret. Therefore, a call to IAccessibleText::caretOffset on the object with focus must always return the HRESULT S_OK and a valid offset value. If the offset value returned indicates an embed character, the client must be able to use the IAccessibleHypertext interface as described in the previous section to downwardly traverse the object hierarchy until it finds the object containing the caret."

The code got mixed together because the end offset is used for the
caret offset. We fix this by only incrementing the end offset
when the selection is being retrieved for selection offsets, as
opposed to for a caret offset.

Fixed: 378692760
Change-Id: I5b66f2cb4ceb57447f0910f24e18e7a4d775ee04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6091642
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Cr-Commit-Position: refs/heads/main@{#1398299}