Refactor: Move extension disable/enable logic to ExtensionRegistrar.
This cl initiates of the migration of extension
disable/enable functionality from `ExtensionService` to
`ExtensionRegistrar`. The goal is to consolidate
extension management within `ExtensionRegistrar` and eventually
remove these methods from `ExtensionService`.

This is the first cl of this refactoring effort, and I would
like to make sure this aligns with the overall refactoring approach
for ExtensionService, and proceed with more followup cls towards this effort in the following steps:

1. Consolidate extension management logic into `ExtensionRegistrar`.

2. Move these methods out of the ExtensionService by making
ExtensionRegistrar a BrowesrContext keyed service (similar
to ExtensionRegistry), so that it does need to accessed
via ExtensionService.

Bug: 40355585
Change-Id: I67bd09a615f0053ae80d3733ae17f8196a62384f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6096354
Reviewed-by: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398336}