Enable secure time by default on desktop
This matches the SecureTimeCertificateVerification.gcl config which
will be set to LAUNCHED after this lands.
There is no need to condition the FetchBehavior param on OS, since it's
default disabled for Android, and once this launches on Android it will
use the same FETCHES_IN_BACKGROUND_AND_ON_DEMAND setting

Metrics Doc: go/securetimemetrics

Tests changed:
-NetworkTimeTrackerTest.NoNetworkQueryWhileFeatureDisabled: Had an check
that TimerDelayForTesting was 0, however this is not set by field_trial_test_ when the feature is set to disabled for testing, so it was relying on the default code behavior. This expectation was removed since the test can still check no query happens without it (through the existing EXPECT_FALSE(tracker_->QueryTimeServiceForTesting()) check).
-SSLErrorClassificationTest.GetClockState: Checks that the right ssl error classification is returned for different clock states. It does not need to run with background time checks enabled (and doing so was causing sequence checker crashes). This configures it so it runs with only on demand checks.


Change-Id: I05bbac4c11bb99761825b56e26040c20cbbef197
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6107312
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Auto-Submit: Carlos IL <carlosil@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398350}