[ios] Use BrowserListObserve API directly
The API of AllWebStateListObservationRegistrar is weird and glitchy
causing subtle issues in all clients (mostly it does not notify or
the Browser being destroyed, and it does not notify of pre-existing
Browsers).

This causes the clients to add work-around, or have bugs. Observing
BrowserList directly is simpler than all those work-arounds, and
each clients has different needs regarding how to deal with the
glitched part of the API.

Convert IOSTranslateInternalsHandler::Observer to directly observe
the BrowserList. Also remove unnecessary complexity regarding the
type of Browsers observed, as it always regular and inactive due
to the call to `ProfileIOS::GetOriginalProfile()` in the existing
implementation.

Bug: 384034363
Change-Id: I1994defb4c9b84d03bfca921f1a6958e404f8565
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6088926
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398478}