[headless] Fix incorrect current screen after Browser.setWindowBounds()
When bounds specified in Browser.setWindowBounds() call result
in window being moved to a different screen, the current screen
reported by (await getScreenDetails()).currentScreen is not
correct.

This CL cleans up headless shell window positioning handling
so that all the necessary notifications are properly sent and
in the correct order.

Drive by: an old workaround for https://crbug.com/40571433
is removed as it is not necessary to artificially bump browser
window size in order for Page.captureScreenshot() to not to
hang after window is moved.

Bug: 381908903, 40571433
Change-Id: I0d7e4ba06b83123b1b9cbf19ecab834a641e811d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6064580
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Peter Kvitek <kvitekp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398488}