Cleanup `RequiresTrustedTypes()`.
We ended up with multiple implementations of a check to determine
whether Trusted Types are enforced, and the spelling of the efficient
implementation was unfortunate.

This patch replaces `IsRequireTrustedTypes()` and
`RequiresTrustedTypes()` with `TrustedTypesRequired()`, and uses the new
method consistently.

No change in behavior is expected. We'll simply use an already-cached
result rather than recalculating it each time.

Change-Id: I9551115d7aef4731b3a03d8faac8422c87a365e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6110393
Commit-Queue: Mike West <mkwst@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398510}