Add metric for filling acceptance grouped by focused field type
A `DenseSet` is used to store for which field types suggestions were
shown.
Another `DenseSet` is used to store for which field types suggestions
were accepted.

When a suggestion is accepted, the corresponding field type is moved
from the first dense set, to the second.

This metric was introduced for the label improvements experiment, but
the metric should be generally useful. Thus, it should be preseverved
beyond the improved labels experiment.

Bug: 380273791
Change-Id: Iad0219cd56d016a27b7867a382afad709aa5bcc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6110212
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Commit-Queue: Luc Nguyen <lucnguyen@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Reviewed-by: Luc Nguyen <lucnguyen@google.com>
Auto-Submit: Theo Cristea <theocristea@google.com>
Cr-Commit-Position: refs/heads/main@{#1398577}