Log SBClientPhishing.NetworkResult even when there is no HTTP response
Previously, this histogram only logged if there was an HTTP response
populated. This means that if there are cases where there is a net error
and no HTTP response, this histogram does not log the net error.

This CL changes the histogram to log whether or not the HTTP response is
populated, matching the functionality of other safe_browsing
NetworkResult histogram logs.

The histogram is renamed to SBClientPhishing.NetworkResult2 due to the
functionality change.

This CL also modifies the histogram owner from drubery@ to andysjlim@.

New histogram:
 - SBClientPhishing.NetworkResult2
Removed histogram:
 - SBClientPhishing.NetworkResult

OBSOLETE_HISTOGRAM[SBClientPhishing.NetworkResult]=Replaced by SBClientPhishing.NetworkResult2.

Bug: 383994656
Change-Id: Ic0cad21f201920b3ae22e53f47e671c755e98e90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6098915
Reviewed-by: Seung Jae Lim <andysjlim@chromium.org>
Commit-Queue: thefrog <thefrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398617}