[Data Sharing] Ensure originalProfile is used
Rather confusingly the initial profile supplied to RootUiCoordinator may
be an OTR profile if the browser was started in incognito mode. It looks
like this wasn't accounted for when starting the Data Sharing service.
Fortunately, asserts are in place to catch this.

Fixed: 384849625
Change-Id: I1ea31f8aa5a8ade25dccf2bce0248f4d54c9bfe5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6108090
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398634}